refactor: pack Primitive and Material into struct
The `Primitive` enum changed its name to `Shape`; New `Primitive` struct contains `Shape` and `Material` and holds `new_...` methods previously present in enum. The change is a foundation of shading system that will be implemented in the next commits.

Minor fix: triangle intersection returns intersection point to match the behavior of other shapes.